requirements.txt
beautifulsoup4
yfinance
requests
snowflake-snowpark-python
snowflake-snowpark-python[pandas]
plotly
streamlit
matplotlib
ta
scikit-learn